Fix issues in pr: delete file with tests, reorder imports, rename packages.
